expanding the wavefront
from all these phases emerges a complete and entirely mechanical kinematic pressurised elasticity dropping in intensity over time as the last scattering moves further and further away. The CMB marks the boundary of the scale of pressure being exerted on each space in location self, with the shadow cones being a linear projected propagation against an occluding object (particle scale) to another occluding object as they each displace their volume of space. There is mounting evidence within the framework itself indicating matter also occludes space from spations, which may require a reformulation of the exact mechanism of the gradiated pressure dynamics, as in my mind I had an incompressible hardpoint being crushed out of its stacking order to fundamentally change the kissing structure, but it might be simply that as we are hard up against earth, they are hard up against nuclei using the same mechanic that pushes the electron toward the nucleus. we shall investigate.